+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - **ID3v2**: The footer flag is written to the tag
+- **ID3v2**: Pictures are written when using `Tag`
 
 ## [0.5.1] - 2022-02-21